+++ docs/changeLog.md
@@ -6,6 +6,16 @@ title: 更新日志
 
 遵循 [Semantic Versioning 2.0.0](https://semver.org/lang/zh-CN/) 语义化版本规范。
 
+### 🚀 2.0.0-bate14
+
+`todo`
+
+- TdUpload
+  - 优化 [onUpload](/high-coupling/td-upload#ref) 回调函数的写法，较上个版本有较多变化，注意项目修改
+  - 优化内置的 request 调用，改为 localConfig 中进行调用
+- tools
+  - 删除方法 downloadBlob，被 [DownloadBlob.use()](/components/download-blob#function) 代替了
+
 ### 🚀 2.0.0-bate13
 
 `2020-11-19`